+++ see-realview/src/main/java/com/see/realview/_core/config/RedisConfig.java
@@ -3,7 +3,7 @@ package com.see.realview._core.config;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.see.realview.domain.token.entity.Token;
+import com.see.realview.token.entity.Token;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
