+++ see-realview/src/main/java/com/see/realview/_core/security/JwtAuthenticationFilter.java
@@ -3,9 +3,9 @@ package com.see.realview._core.security;
 import com.see.realview._core.exception.client.BadRequestException;
 import com.see.realview._core.exception.ExceptionStatus;
 import com.see.realview._core.exception.client.UnauthorizedException;
-import com.see.realview.domain.token.entity.Token;
-import com.see.realview.domain.token.service.TokenServiceImpl;
-import com.see.realview.domain.user.entity.UserAccount;
+import com.see.realview.token.entity.Token;
+import com.see.realview.token.service.TokenServiceImpl;
+import com.see.realview.user.entity.UserAccount;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
